+++ cli/tests/sous_test.go
@@ -3,6 +3,10 @@ package tests
 import (
 	"log"
 	"testing"
+
+	"github.com/opentable/sous/config"
+	"github.com/opentable/sous/graph"
+	"github.com/opentable/sous/util/yaml"
 )
 
 func TestSous(t *testing.T) {
@@ -29,4 +33,57 @@ func TestSousVersion(t *testing.T) {
 	term.Stderr.ShouldHaveNumLines(0)
 	term.Stdout.ShouldHaveNumLines(1)
 	term.Stdout.ShouldHaveLineContaining("sous version")
+}
+
+func TestSousConfig_validConfig(t *testing.T) {
+	term := NewTerminal(t, "0.0.0")
+
+	defer term.PrintFailureSummary()
+
+	testConfig := &config.Config{}
+	if err := term.Graph.Realise(testConfig); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := testConfig.Validate(); err != nil {
+		t.Fatalf("inconclusive; config was not valid to begin with: %s", err)
+	}
+
+	term.RunCommand("sous config")
+
+	yamlConfig, err := yaml.Marshal(testConfig)
+	if err != nil {
+		t.Fatalf("inconclusive: unable to marshal config to YAML: %s", err)
+	}
+	term.Stdout.ShouldContain(yamlConfig)
+
+	term.Stderr.ShouldBeEmpty()
+}
+
+func TestSousConfig_invalidConfig(t *testing.T) {
+	term := NewTerminal(t, "0.0.0")
+
+	defer term.PrintFailureSummary()
+
+	testConfig := graph.RawConfig{}
+	if err := term.Graph.Realise(&testConfig); err != nil {
+		t.Fatal(err)
+	}
+	testConfig.Server = "not a valid URL"
+	if err := testConfig.Validate(); err == nil {
+		t.Fatalf("inconclusive; config was not invalid")
+	}
+	// Put the invalid config back in the graph.
+	term.Graph.Replace(testConfig)
+
+	term.RunCommand("sous config")
+
+	yamlConfig, err := yaml.Marshal(testConfig.Config)
+	if err != nil {
+		t.Fatalf("inconclusive: unable to marshal config to YAML: %s", err)
+	}
+	term.Stdout.ShouldContain(yamlConfig)
+
+	const configWarning = `WARNING: Invalid configuration: Config.Server: URL "not a valid URL" must begin with http:// or https://`
+	term.Stderr.ShouldContainString(configWarning)
 }