+++ test/server_test.go
@@ -61,7 +61,7 @@ func (suite serverTests) prepare(extras ...interface{}) http.Handler {
 	g := graph.TestGraphWithConfig(semv.Version{}, &bytes.Buffer{}, os.Stdout, os.Stdout,
 		"StateLocation: '"+outpath+"'\n")
 	g.Add(extras...)
-	g.Add(&config.Verbosity{})
+	g.Add(graph.VerbosityOverride{})
 	g.Add(&config.DeployFilterFlags{})
 	g.Add(graph.DryrunBoth)